change: AbstractController.java         AbstractService.java         Adottante.java         AdottanteController.java         AdottanteDto.java         AdottanteRepository.java         AdottanteService.java         AdozioneRequestDto.java         Animale.java         AnimaleController.java         AnimaleDto.java         AnimaleRepository.java         AnimaleService.java         CentroAdozione.java         CentroAdozioneController.java         CentroAdozioneDto.java         CentroAdozioneRepository.java         CentroAdozioneService.java         PraticaAdozione.java         PraticaAdozioneController.java         PraticaAdozioneDto.java         PraticaAdozioneRepository.java         PraticaAdozioneService.java         ServiceDTO.java         VisitaMedicaController.java         VisitaMedicaDto.java         VisitaMedicaRepository.java         Volontario.java         VolontarioController.java         VolontarioDto.java         VolontarioRepository.java

+++ animali/animali/src/main/java/it/adozioni/animali/Controller/AbstractController.java
@@ -17,7 +17,7 @@ public abstract class AbstractController <DTO>{
     private ServiceDTO<DTO> service;
 
     @DeleteMapping("/delete")
-    public void delete(@RequestParam("id") Integer id) {
+    public void delete(@RequestParam("id") Long id) {
         service.delete(id);
     }
 
@@ -35,7 +35,7 @@ public abstract class AbstractController <DTO>{
     }
 
     @GetMapping("/read")
-    public DTO read(@RequestParam("id") Integer id) {
+    public DTO read(@RequestParam("id") Long id) {
         return service.read(id);
     }
