+++ animali/animali/src/main/java/it/adozioni/animali/Repository/PraticaAdozioneRepository.java
@@ -5,11 +5,11 @@ import it.adozioni.animali.Model.StatoPratica;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface PraticaAdozioneRepository extends JpaRepository<PraticaAdozione, Integer> {
-    List<PraticaAdozione> findByAdottanteId(Integer adottanteId);
+public interface PraticaAdozioneRepository extends JpaRepository<PraticaAdozione, Long> {
+    List<PraticaAdozione> findByAdottanteId(Long adottanteId);
 
     // Utile per controllare se esiste già una pratica aperta per lo stesso animale dallo stesso utente
-    boolean existsByAdottanteIdAndAnimaleIdAndStatoNot(Integer adottanteId, Integer animaleId, StatoPratica stato);
+    boolean existsByAdottanteIdAndAnimaleIdAndStatoNot(Long adottanteId, Long animaleId, StatoPratica stato);
 
 
 }
