+++ animali/animali/src/main/java/it/adozioni/animali/Service/ServiceDTO.java
@@ -7,12 +7,12 @@ public interface ServiceDTO<DTO> {
 
     public Iterable<DTO> getAll();
 
-    public DTO read(Integer id);
+    public DTO read(Long id);
 
     public DTO insert (DTO dto);
 
     public DTO update (DTO dto);
 
-    public void delete (Integer id);
+    public void delete (Long id);
 }
 //
